Finish Day06 part 2

+++ source/Day06/Lanternfish1.cs
@@ -4,7 +4,7 @@ namespace AdventOfCode2021.Day06
 {
 	public static class Lanternfish1
 	{
-		public static int Run(string[] lines)
+		public static long Run(string[] lines)
 		{
 			var startingLanternfishTimers = lines[0].Split(',').Select(x => int.Parse(x));
 			var lanternfishManager = new LanternfishManager(startingLanternfishTimers);
